% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apertium_get_language_pairs.R
\name{apertium_get_language_pairs}
\alias{apertium_get_language_pairs}
\title{Get Apertium Language Pairs}
\usage{
apertium_get_language_pairs(host = "https://apertium.org/apy")
}
\arguments{
\item{host}{Host URL for the Apertium API (default is "https://apertium.org/apy").}
}
\value{
A list of language pairs. Each element contains sourceLanguage and targetLanguage.
}
\description{
This function retrieves the supported language pairs from the Apertium API.
}
\examples{
\donttest{
pairs <- apertium_get_language_pairs()
head(pairs, 5)

}
}
