% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_translation_table.R
\name{create_translation_table}
\alias{create_translation_table}
\title{Create a Translation Table}
\usage{
create_translation_table(words, languages)
}
\arguments{
\item{words}{A character vector containing the words to be translated.}

\item{languages}{A character vector specifying the target languages for translation.}
}
\value{
A data frame representing the translation table with original words and translations in each language.
}
\description{
This function generates a translation table by translating a list of words into multiple languages.
}
\examples{
\dontrun{
words <- c("Hello", "Translate", "Table", "Script")
languages <- c("es", "fr", "de", "nl")
translations <- create_translation_table(words, languages)
print(translations)
}
}
