% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyglot.R
\name{fib_assemblyscript}
\alias{fib_assemblyscript}
\title{Fibonacci number in Assemblyscript}
\usage{
fib_assemblyscript(n)
}
\arguments{
\item{n}{an integer >= 1}
}
\description{
Fibonacci number in Assemblyscript
}
