% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_is_valid_language_code.R
\name{google_is_valid_language_code}
\alias{google_is_valid_language_code}
\title{Check if a language code is valid}
\usage{
google_is_valid_language_code(language_code)
}
\arguments{
\item{language_code}{The language code to check.}
}
\value{
A logical value indicating if the language code is valid.
}
\description{
This function checks if a given language code is in the \code{google_supported_languages} dataset.
}
\examples{
google_is_valid_language_code("en") # TRUE
google_is_valid_language_code("fr") # TRUE
google_is_valid_language_code("xx") # FALSE
}
