% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microsoft_supported_languages.R
\name{microsoft_supported_languages}
\alias{microsoft_supported_languages}
\title{Get the set of languages currently supported by the Microsoft Translator API}
\usage{
microsoft_supported_languages(scope = NULL)
}
\arguments{
\item{scope}{(optional) A comma-separated list of names defining the group of languages to return. Allowed group names are: translation, transliteration, and dictionary. If no scope is given, then all groups are returned.}
}
\value{
A list of supported languages for the specified groups.
}
\description{
Get the set of languages currently supported by the Microsoft Translator API
}
\examples{
\dontrun{
microsoft_supported_languages(scope = "translation,transliteration,dictionary")
}
}
