% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymemory_translate.R
\name{mymemory_translate}
\alias{mymemory_translate}
\title{Translate text using mymemory translate}
\usage{
mymemory_translate(text, target_language = "en", source_language = "auto")
}
\arguments{
\item{text}{Text to translate.}

\item{target_language}{Language to translate text to.}

\item{source_language}{Language to translate text from}
}
\value{
Translated text.
}
\description{
Translate text using mymemory translate
}
\examples{
\dontrun{
mymemory_translate("Hello World", target_language = "es", source_language = "en")
}
}
