% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pons_dictionaries.R
\name{pons_dictionaries}
\alias{pons_dictionaries}
\title{Get the list of available dictionaries from PONS API}
\usage{
pons_dictionaries(language = "en")
}
\arguments{
\item{language}{The language of the output (ISO 639-1 - two-letter codes). Supported languages are de, el, en, es, fr, it, pl, pt, ru, sl, tr, zh.}
}
\value{
A list of available dictionaries in the specified language.
}
\description{
Get the list of available dictionaries from PONS API
}
\examples{
\dontrun{
pons_dictionaries(language = "es")
}
}
