% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pons_word_definitions.R
\name{pons_word_definitions}
\alias{pons_word_definitions}
\title{Get definitions for words from PONS dictionaries}
\usage{
pons_word_definitions(word, source_language, target_language, limit = 5)
}
\arguments{
\item{word}{The word to get definitions for.}

\item{source_language}{The source language code (ISO 639-1 - two-letter codes).}

\item{target_language}{The target language code (ISO 639-1 - two-letter codes).}

\item{limit}{Maximum number of definitions to return. Default is 5.}
}
\value{
A data frame with word definitions and grammatical information.
}
\description{
Get definitions for words from PONS dictionaries
}
\examples{
\dontrun{
pons_word_definitions("hello", "en", "es")
pons_word_definitions("casa", "es", "en", limit = 3)
}
}