% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pons_word_synonyms.R
\name{pons_word_synonyms}
\alias{pons_word_synonyms}
\title{Get synonyms for words from PONS dictionaries}
\usage{
pons_word_synonyms(word, source_language, target_language, limit = 5)
}
\arguments{
\item{word}{The word to get synonyms for.}

\item{source_language}{The source language code (ISO 639-1 - two-letter codes).}

\item{target_language}{The target language code (ISO 639-1 - two-letter codes).}

\item{limit}{Maximum number of synonyms to return. Default is 5.}
}
\value{
A data frame with synonyms and related words.
}
\description{
Get synonyms for words from PONS dictionaries
}
\examples{
\dontrun{
pons_word_synonyms("hello", "en", "es")
pons_word_synonyms("casa", "es", "en", limit = 3)
}
}