% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcri_api_key.R
\name{qcri_api_key}
\alias{qcri_api_key}
\title{Get the QCRI API key from the environment variable}
\usage{
qcri_api_key()
}
\value{
The QCRI API key stored in the QCRI_API_KEY environment variable.
}
\description{
Get the QCRI API key from the environment variable
}
