% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcri_get_language_pairs.R
\name{qcri_get_language_pairs}
\alias{qcri_get_language_pairs}
\title{QCRI Get Language Pairs}
\usage{
qcri_get_language_pairs(api_key = qcri_api_key())
}
\arguments{
\item{api_key}{The API key associated with the user account being used. If not provided, the function will attempt to retrieve it from the QCRI_API_KEY environment variable. You can register for an API key at https://mt.qcri.org/api/register}
}
\value{
Language pairs.
}
\description{
This function retrieves the supported language pairs from the QCRI Multiterm API.
}
\examples{
\dontrun{
qcri_get_language_pairs(api_key = "YourApiKey")
qcri_get_language_pairs()
}
}
