% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_apertium.R
\name{translate_apertium}
\alias{translate_apertium}
\title{Translate text using Apertium}
\usage{
translate_apertium(text, target_language, source_language, host = "https://apertium.org/apy")
}
\arguments{
\item{text}{Text to translate. Can be a single string or a vector of strings.}

\item{target_language}{Language to translate text to.}

\item{source_language}{Language to translate text from.}

\item{host}{Host URL for the Apertium API (default is "https://apertium.org/apy").}
}
\value{
Translated text. Returns a vector if input is a vector.
}
\description{
Translate text using Apertium
}
\examples{
\donttest{
translate_apertium("Hello World", target_language = "es", source_language = "en")
translate_apertium("Hola mundo", target_language = "en", source_language = "es")

# Translate multiple texts
texts <- c("Hello", "Good morning", "Thank you")
translate_apertium(texts, target_language = "es", source_language = "en")
}
}