% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_to_morse.R
\name{translate_to_morse}
\alias{translate_to_morse}
\title{Translate Text to Morse Code using the FunTranslations API}
\usage{
translate_to_morse(text, api_key = NULL)
}
\arguments{
\item{text}{A character string containing the text to be translated to Morse code.}

\item{api_key}{(optional) Your FunTranslations API key, if you have a paid subscription.}
}
\value{
A list containing the translated Morse code text and other metadata.
}
\description{
This function takes a string of text as input and translates it to Morse code using the FunTranslations API.
}
