% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikimedia_detect_language.R
\name{wikimedia_detect_language}
\alias{wikimedia_detect_language}
\title{Detect the language of a text}
\usage{
wikimedia_detect_language(text)
}
\arguments{
\item{text}{The text whose language is to be detected.}
}
\value{
The detected language.
}
\description{
This function sends a POST request to the Wikimedia Language ID API with the specified text,
parses the JSON response, and returns the detected language.
}
\examples{
{
\donttest{
# Detect the language of a text
wikimedia_detect_language("Hallo, wereld")
}
}
}
