% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikipedia_get_language_names.R
\name{wikipedia_get_language_names}
\alias{wikipedia_get_language_names}
\title{Get language names}
\usage{
wikipedia_get_language_names()
}
\value{
A dataframe of language names.
}
\description{
This function sends a GET request to the Wikipedia API and returns the language names as a dataframe.
}
\examples{
# Get language names
wikipedia_get_language_names()

}
