% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmcloud_translate.R
\name{wmcloud_translate}
\alias{wmcloud_translate}
\title{Translate text using WMCloud}
\usage{
wmcloud_translate(
  text,
  target_language = "en",
  source_language = "en",
  format = "text",
  model = "nllb200-600M"
)
}
\arguments{
\item{text}{The text to translate.}

\item{target_language}{The target language for the translation (default is "en").}

\item{source_language}{The source language of the text (default is "en").}

\item{format}{The format of the content ("html", "json", "markdown", "text", "svg", "webpage").}

\item{model}{The model to use for the translation (only "nllb200-600M" is currently known to work).}
}
\value{
The translated text.
}
\description{
This function sends a POST request to the WMCloud translation API with the specified parameters,
parses the JSON response, and returns the translated text.
}
