% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_translate.R
\name{batch_translate}
\alias{batch_translate}
\title{Batch Translation Function}
\usage{
batch_translate(input_file, source_language, target_languages)
}
\arguments{
\item{input_file}{A character string indicating the path to the input file.}

\item{source_language}{A character string indicating the source language.}

\item{target_languages}{A character vector indicating the target languages.}

\item{output_dir}{A character string indicating the directory for the output file. If not provided, the output file is saved in the current working directory.}
}
\value{
Nothing is returned.
}
\description{
This function translates a file into each target language using the polyglotr package's translate_file function, and saves the translated files in the specified directory.
}
\examples{
\dontrun{
batch_translate("README.md", "nl", c("fr", "es", "de"))
}
}
