% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_get_supported_languages.R
\name{google_get_supported_languages}
\alias{google_get_supported_languages}
\title{Get Supported Languages}
\usage{
google_get_supported_languages()
}
\value{
A data frame containing the supported languages and their corresponding ISO 639-1 codes.
}
\description{
This function fetches the supported languages from the Google Cloud Translate documentation page.
}
