% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_supported_languages.R
\docType{data}
\name{google_supported_languages}
\alias{google_supported_languages}
\title{Google Supported Languages}
\format{
A data frame with two variables: language_name and iso_code
}
\source{
Google Translate API
}
\usage{
google_supported_languages
}
\description{
This dataset contains the language names and iso codes of languages supported by Google Translate API.
}
\keyword{data}
