% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_translate.R
\name{google_translate}
\alias{google_translate}
\title{Translate text using google translate}
\usage{
google_translate(text, target_language = "en", source_language = "auto")
}
\arguments{
\item{text}{Text to translate.}

\item{target_language}{Language to translate text to.}

\item{source_language}{Language to translate text from}
}
\value{
Translated text.
}
\description{
Translate text using google translate
}
\examples{
\dontrun{
google_translate("I love languages", target_language = "es")
}
}
