% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_translate.R
\name{google_translate}
\alias{google_translate}
\title{Translate text using google translate}
\usage{
google_translate(text, target_language = "en", source_language = "auto")
}
\arguments{
\item{text}{This is the text that you want to translate.}

\item{target_language}{This is the language that you want to translate the text into.
The default value for this argument is "en" for English.}

\item{source_language}{This is the language of the text that you want to translate.
The default value for this argument is "auto",
which means that the function will try to automatically detect the language of the text.}
}
\value{
Translated text.
}
\description{
Translate text using google translate
}
\examples{
\donttest{
google_translate("I love languages", target_language = "es")
text_to_translate <- c("the", "quick", "brown")
google_translate (text_to_translate, "fr", "en")
}
}
