% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language_detect.R
\name{language_detect}
\alias{language_detect}
\title{Detect Language using Google Translate API}
\usage{
language_detect(text)
}
\arguments{
\item{text}{The text for which the language needs to be detected.}
}
\value{
A character string representing the detected language.
}
\description{
This function detects the language of a given text using the Google Translate API.
}
