% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_translate_file.R
\name{translate_file}
\alias{translate_file}
\title{Translate File}
\usage{
translate_file(
  file_path,
  target_language = "en",
  source_language = "auto",
  overwrite = FALSE
)
}
\arguments{
\item{file_path}{The path to the file to be translated.}

\item{target_language}{The target language to translate the file content to. Default is "en".}

\item{source_language}{The source language of the file content. Default is "auto".}

\item{overwrite}{Logical indicating whether to overwrite the original file with the translated content. Default is FALSE.}
}
\description{
Translates the content of a file using Google Translate API.
}
\examples{
\dontrun{
translate_file("path/to/file.txt", target_language = "fr", source_language = "en", overwrite = TRUE)
}
}
